% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_my_packages.R
\name{build_my_packages}
\alias{build_my_packages}
\title{Build my packages}
\usage{
build_my_packages(..., package_path = PKG_PATH)
}
\arguments{
\item{...}{package names as characters}
}
\description{
Build my packages
}
